+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Updated: Race times for the expansion 'Dragonflight'
-- Updated: ruRU localization (#2 - Thanks to Hollicsh)
+- Added: New Race Tracker backgrounds (Alliance and Horde)
+- Changed: The speed display during a race is shown as a bar
+- Minor code adjustments